+++ qbraid/runtime/quantinuum/__init__.py
@@ -1,4 +1,4 @@
-# Copyright 2025 qBraid
+# Copyright 2026 qBraid
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -13,22 +13,27 @@
 # limitations under the License.
 
 """
-PyTKET conversions
+Module for submitting and managing jobs through the Quantinuum NEXUS API.
 
-.. currentmodule:: qbraid.transpiler.conversions.pytket
+.. currentmodule:: qbraid.runtime.quantinuum
 
-Functions
-----------
+Classes
+--------
 
 .. autosummary::
    :toctree: ../stubs/
 
-   pytket_to_braket
-   pytket_to_qasm2
-   pytket_to_qiskit
+    QuantinuumProvider
+    QuantinuumDevice
+    QuantinuumJob
 
 """
-from .pytket_extras import pytket_to_braket, pytket_to_qiskit
-from .pytket_to_qasm2 import pytket_to_qasm2
+from .device import QuantinuumDevice
+from .job import QuantinuumJob
+from .provider import QuantinuumProvider
 
-__all__ = ["pytket_to_braket", "pytket_to_qasm2", "pytket_to_qiskit"]
+__all__ = [
+    "QuantinuumDevice",
+    "QuantinuumJob",
+    "QuantinuumProvider",
+]