+++ tests/runtime/rigetti/__init__.py
@@ -1,4 +1,4 @@
-# Copyright 2025 qBraid
+# Copyright 2026 qBraid
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -11,24 +11,3 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-
-"""
-PyTKET conversions
-
-.. currentmodule:: qbraid.transpiler.conversions.pytket
-
-Functions
-----------
-
-.. autosummary::
-   :toctree: ../stubs/
-
-   pytket_to_braket
-   pytket_to_qasm2
-   pytket_to_qiskit
-
-"""
-from .pytket_extras import pytket_to_braket, pytket_to_qiskit
-from .pytket_to_qasm2 import pytket_to_qasm2
-
-__all__ = ["pytket_to_braket", "pytket_to_qasm2", "pytket_to_qiskit"]